+++ services/CHANGELOG.md
@@ -11,6 +11,10 @@
 
 **New Features**
 
+* LocalSocketProtocol: a replacement for SpeakEasy.
+* ShellCommandLocalSocketClient: client that speaks LocalSocketProtocol.
+* ShellCommandLocalSocketExecutorServer: server that speaks LocalSocketProtocol.
+
 **Breaking Changes**
 
 **API Changes**